# src/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "My Cookery Book 2 - backend",
        "version": "0.0.2",
        "license": {
            "name": "MIT"
        },
        "description": "This is rest api application backend developed on NodeJS and ExpressJS. Writen in typescript.",
        "contact": {
            "email": "tfilosk@gmail.com",
            "name": "Tomáš Filo",
            "url": "https://github.com/tfilo/my-cookery-book-2"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development"
        }
    ],
...
    "paths": {
...
        "/unit": {
            "post": {
                "operationId": "createUnit",
                "tags": ["unit"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUnit"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Unit created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unit"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "409": {
                        "$ref": "#/components/responses/ConstraintError"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/unit/byUnitCategory/{unitCategoryId}": {
            "get": {
                "operationId": "getUnitsByUnitCategory",
                "tags": ["unit"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "unitCategoryId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of units by unit category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SimpleUnit"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/unit/{unitId}": {
            "get": {
                "operationId": "getUnit",
                "tags": ["unit"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "unitId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Unit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unit"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "put": {
                "operationId": "updateUnit",
                "tags": ["unit"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "unitId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUnit"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Unit updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unit"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/ConstraintError"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "delete": {
                "operationId": "deleteUnit",
                "tags": ["unit"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "unitId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Unit deleted"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/unitCategory": {
            "post": {
                "operationId": "createUnitCategory",
                "tags": ["unitCategory"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUnitCategory"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Unit category created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnitCategory"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "409": {
                        "$ref": "#/components/responses/ConstraintError"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "get": {
                "operationId": "getUnitCategories",
                "tags": ["unitCategory"],
                "responses": {
                    "200": {
                        "description": "List of unit categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SimpleUnitCategory"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
        "/unitCategory/{unitCategoryId}": {
            "get": {
                "operationId": "getUnitCategory",
                "tags": ["unitCategory"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "unitCategoryId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "UnitCategory",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnitCategory"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "put": {
                "operationId": "updateUnitCategory",
                "tags": ["unitCategory"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "unitCategoryId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUnitCategory"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Unit category updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnitCategory"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "409": {
                        "$ref": "#/components/responses/ConstraintError"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            },
            "delete": {
                "operationId": "deleteUnitCategory",
                "tags": ["unitCategory"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "unitCategoryId",
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "minimum": 1
                        },
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Unit category deleted"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        },
        "schemas": {
...
            "ErrorTypeEnum": {
                "type": "string",
                "enum": [
                    "GENERAL_ERROR",
                    "DATABASE_ERROR",
                    "NOT_FOUND",
                    "VALIDATION_FAILED",
                    "INVALID_CREDENTIALS",
                    "FORBIDEN",
                    "EXPIRED_TOKEN",
                    "INVALID_TOKEN",
                    "UNIQUE_CONSTRAINT_ERROR",
                    "UNABLE_TO_SENT_EMAIL",
                    "CONSTRAINT_FAILED",
                    "TOO_MANY_REQUESTS"
                ]
            },
...
            "UnitCategory": {
                "type": "object",
                "required": ["id", "name", "createdAt", "updatedAt"],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-04-10T12:58:33.592Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-04-10T12:58:33.592Z"
                    }
                }
            },
            "SimpleUnitCategory": {
                "type": "object",
                "required": ["id", "name"],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                    }
                }
            },
            "CreateUnitCategory": {
                "type": "object",
                "required": ["name"],
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                    }
                }
            },
            "UpdateUnitCategory": {
                "type": "object",
                "$ref": "#/components/schemas/CreateUnitCategory"
            },
            "Unit": {
                "type": "object",
                "required": ["id", "name", "abbreviation", "required", "unitCategoryId", "createdAt", "updatedAt"],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                    },
                    "abbreviation": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 20
                    },
                    "required": {
                        "type": "boolean"
                    },
                    "unitCategoryId": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-04-10T12:58:33.592Z"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-04-10T12:58:33.592Z"
                    }
                }
            },
            "SimpleUnit": {
                "type": "object",
                "required": ["id", "name", "abbreviation", "required"],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                    },
                    "abbreviation": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 20
                    },
                    "required": {
                        "type": "boolean"
                    }
                }
            },
            "CreateUnit": {
                "type": "object",
                "required": ["name", "abbreviation", "required", "unitCategoryId"],
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80
                    },
                    "abbreviation": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 20
                    },
                    "required": {
                        "type": "boolean"
                    },
                    "unitCategoryId": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    }
                }
            },
            "UpdateUnit": {
                "type": "object",
                "$ref": "#/components/schemas/CreateUnit"
            },
...
            "Error": {
                "type": "object",
                "required": ["code"],
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "code": {
                        "$ref": "#/components/schemas/ErrorTypeEnum"
                    },
                    "fields": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "responses": {
            "UnauthorizedError": {
                "description": "Access token is missing or invalid. Possibly incorrect credentials.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Object is missing, or required parent doesn't exists.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "InternalServerError": {
                "description": "General error. Something went wrong on server side.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
            "UnprocessableEntity": {
                "description": "Request can't be processed because validation failed.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
            "ConstraintError": {
                "description": "Constraint or duplicity error.",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            },
...
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}